+++ src/main/scala/org/clulab/asist/agents/package.scala
@@ -1,0 +1,3 @@
+package org.clulab.asist package object agents {
+
+}
