+++ src/main/scala/org/clulab/asist/extraction/TomcatActions.scala
@@ -3,14 +3,11 @@ package org.clulab.asist
 
 import com.typesafe.scalalogging.LazyLogging
 import org.clulab.odin._
-import org.clulab.asist.AsistEngine._
+import org.clulab.asist.TomcatRuleEngine._
 
 import scala.collection.mutable.ArrayBuffer
 
-class TomcatActions(
-    val timeintervals: (ArrayBuffer[Int], ArrayBuffer[Int], ArrayBuffer[Int])
-) extends Actions
-    with LazyLogging {
+class TomcatActions() extends Actions with LazyLogging {
 
   def passThrough(
       mentions: Seq[Mention],
@@ -59,39 +56,6 @@ class TomcatActions(
     localState.lookUpTable.values.toStream.flatten.distinct.toVector
   }
 
-  def removeResearcher(
-      mentions: Seq[Mention],
-      state: State = new State()
-  ): Seq[Mention] = {
-    // If there are no timeintervals, return all events found
-    if (timeintervals._1.size == 0) {
-      return mentions
-    }
-    val to_be_returned = new ArrayBuffer[Mention]
-    for (men <- mentions) {
-      val startOffset = men match {
-        case cur: EventMention     => cur.trigger.startOffset
-        case cur: TextBoundMention => cur.startOffset
-      }
-
-      var cur = timeintervals._1(0)
-      var i = 0
-      while (cur < startOffset) {
-        i += 1
-        cur = timeintervals._1(i)
-      }
-      if (!(timeintervals._2 contains i)) {
-        to_be_returned.append(men)
-        if (timeintervals._3 contains timeintervals._1(i - 1)) {
-          // This checks if the previous utterance was a researcher question
-          //println(men.words.mkString)
-          // TODO create new event mention
-        }
-      }
-    }
-    to_be_returned
-  }
-
   def mkVictim(mentions: Seq[Mention], state: State = new State()): Seq[Mention] = {
     mentions.map(addArgLabel(_, TARGET_ARG, VICTIM, Some(ENTITY)))
   }
@@ -134,12 +98,6 @@ class TomcatActions(
 }
 
 object TomcatActions {
-
-  def apply(
-      timeintervals: (ArrayBuffer[Int], ArrayBuffer[Int], ArrayBuffer[Int])
-  ) =
-    new TomcatActions(
-      timeintervals: (ArrayBuffer[Int], ArrayBuffer[Int], ArrayBuffer[Int])
-    )
+  def apply() = new TomcatActions()
 }
