+++ src/main/scala/org/clulab/asist/package.scala
@@ -1,0 +1,3 @@
+package org.clulab package object asist {
+
+}
