+++ src/test/scala/org/clulab/asist/rules/BaseConceptTest.scala
