+++ src/test/scala/org/clulab/asist/rules/CommitmentTest.scala
