+++ src/test/scala/org/clulab/asist/rules/EntityTest.scala
