+++ src/test/scala/org/clulab/asist/rules/TestCommunications.scala
@@ -11,7 +11,7 @@ class TestCommunications extends BaseTest {
   passingTest should "Parse existential constructions" in {
     val text = "There is a medkit in the library. "
 
-    val mentions = extractor.extractFromText(text)
+    val mentions = extractor.extractFrom(text)
     val medkit = DesiredMention("MedKit", "medkit")
     val location = DesiredMention("Infrastructure", "library")
     val ex1 = DesiredMention("KnowledgeSharing", "There is a medkit in the library", Map("location" -> Seq(location), "exists" -> Seq(medkit)))
@@ -23,7 +23,7 @@ class TestCommunications extends BaseTest {
   failingTest should "Parse existential constructions 2" in {
     val text =  "There is a victim in here."
     // fixme:{I dont know why this one fails! Help}
-    val mentions = extractor.extractFromText(text)
+    val mentions = extractor.extractFrom(text)
     val deic = DesiredMention("Deictic", "here")
     val victim_men = DesiredMention("Victim", "victim")
     val ex2 = DesiredMention("KnowledgeSharing", "There is a victim in here", Map("location" -> Seq(deic), "exists" -> Seq(victim_men)))
