+++ src/test/scala/org/clulab/asist/rules/TestPrimitives.scala
