+++ src/test/scala/org/clulab/asist/rules/TestQuestions.scala
@@ -15,7 +15,7 @@ class TestQuestions extends BaseTest {
       "Where are they? " +
       "Where is it? " +
       "Where in the back room is it?"
-    val mentions = extractor.extractFromText(text)
+    val mentions = extractor.extractFrom(text)
     val victim = DesiredMention("Victim", "victim")
     val one = DesiredMention("Entity", "one")
     val others = DesiredMention("Entity", "others")
@@ -42,7 +42,7 @@ class TestQuestions extends BaseTest {
 
   it should "find location questions with `location_moving_question` rule" in {
     val text = "Where did they go?"
-    val mentions = extractor.extractFromText(text)
+    val mentions = extractor.extractFrom(text)
     val they = DesiredMention("Entity", "they")
     val move = DesiredMention("Move", "they go", Map("person" -> Seq(they)))
     val desired1 = DesiredMention("LocationQuestion", "Where did they go", Map("topic" -> Seq(move)))
@@ -56,7 +56,7 @@ class TestQuestions extends BaseTest {
       "What's that over there? " +
       "What is the plan? " +
       "What is that?"
-    val mentions = extractor.extractFromText(text)
+    val mentions = extractor.extractFrom(text)
 
     val that = DesiredMention("DemPron", "that")
     val there = DesiredMention("Deictic", "there")
